POM Edit - Dashboard subpackage created, methods and locators separated
